Refactoring JS files: home page

+++ client/js/book.js
@@ -30,7 +30,7 @@ export const getBooks = (endipoint, fun) => {
     .then(error => console.log(error));
 }
 
-export const borrowBook = (book) => {
+const borrowBook = (book) => {
     fetch(`http://127.0.0.1:5000/notification/send-notification?owner_id=${book.owner_id}&book_id=${book.book_id}`, {
         method: 'POST',
         credentials: 'include'
@@ -120,19 +120,26 @@ const createInfoContainer = (book) => {
     return container;
 }
 
-const createActionBtns = (book, bookElement) => {
-    const btns = document.createElement("div");
-    btns.classList.add("action-btns");
-
-    const editBtn = createBtn("Edit", "edit-btn", "edit-btn");
-    editBtn.addEventListener("click", () => handleUpdate(book));
-
-    const deleteBtn = createBtn(`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path d="M135.2 17.7C140.6 6.8 151.7 0 163.8 0L284.2 0c12.1 0 23.2 6.8 28.6 17.7L320 32l96 0c17.7 0 32 14.3 32 32s-14.3 32-32 32L32 96C14.3 96 0 81.7 0 64S14.3 32 32 32l96 0 7.2-14.3zM32 128l384 0 0 320c0 35.3-28.7 64-64 64L96 512c-35.3 0-64-28.7-64-64l0-320zm96 64c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16zm96 0c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16zm96 0c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16z"/></svg>`,
-                                "delete-btn", "delete-btn");
-    deleteBtn.addEventListener("click", () => deleteBook(book.book_id, bookElement));
-
-    btns.append(editBtn, deleteBtn);
-    return btns;
+const createActionBtns = (book, bookElement, user_id) => {
+    if (book.owner_id == user_id) {
+        const btns = document.createElement("div");
+        btns.classList.add("action-btns");
+    
+        const editBtn = createBtn("Edit", "edit-btn", "edit-btn");
+        editBtn.addEventListener("click", () => handleUpdate(book));
+    
+        const deleteBtn = createBtn(`<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path d="M135.2 17.7C140.6 6.8 151.7 0 163.8 0L284.2 0c12.1 0 23.2 6.8 28.6 17.7L320 32l96 0c17.7 0 32 14.3 32 32s-14.3 32-32 32L32 96C14.3 96 0 81.7 0 64S14.3 32 32 32l96 0 7.2-14.3zM32 128l384 0 0 320c0 35.3-28.7 64-64 64L96 512c-35.3 0-64-28.7-64-64l0-320zm96 64c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16zm96 0c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16zm96 0c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16z"/></svg>`,
+                                    "delete-btn", "delete-btn");
+        deleteBtn.addEventListener("click", () => deleteBook(book.book_id, bookElement));
+    
+        btns.append(editBtn, deleteBtn);
+        return btns;
+    }
+    else {
+        const borrowBtn = createBtn("Borrow", "borrow-btn", "borrow-btn");
+        borrowBtn.addEventListener("click", () => borrowBook(book));
+        return borrowBtn;
+    }
 }
 
 const createBorrowBtn = (book) => {
@@ -150,9 +157,14 @@ const createBookElement = (book, index) => {
 
     const infoContainer = createInfoContainer(book);
     const actionBtns = createActionBtns(book, bookElement);
-    const borrowBtn = createBorrowBtn(book);
 
-    bookElement.append(image, infoContainer, actionBtns, borrowBtn);
+    bookElement.append(image, infoContainer, actionBtns);
+
+    const user_id = JSON.parse(localStorage.getItem("user")).user_id;
+    if (book.owner_id == user_id) {
+        const borrowBtn = createBorrowBtn(book);
+        bookElement.appendChild(borrowBtn);
+    }
     return bookElement;
 }
